+++ code/elife/user-server/src/main/java/user/controller/FriendController.java
@@ -23,13 +23,37 @@ public class FriendController {
 
     @RequestMapping(path = "/friendList")
     @ResponseBody
-    public JSONArray friendList(@RequestParam String username){
+    public JSONArray friendList(HttpServletRequest request, @RequestParam String username){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        System.out.println(name);
+        System.out.println(username);
+        System.out.println(role);
+        System.out.println(name.equals(username));
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
+            JSONObject jsonObject = new JSONObject();
+            JSONArray jsonArray = new JSONArray();
+            jsonObject.put("login", 0);
+            jsonArray.add(jsonObject);
+            return jsonArray;
+        }
         return friendService.friendList(username);
     }
 
     @RequestMapping(path = "/friendSearchList")
     @ResponseBody
-    public JSONArray friendSearchList(@RequestParam String username){
+    public JSONArray friendSearchList(HttpServletRequest request, @RequestParam String username){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role)) {
+            JSONObject jsonObject = new JSONObject();
+            JSONArray jsonArray = new JSONArray();
+            jsonObject.put("login", 0);
+            jsonArray.add(jsonObject);
+            return jsonArray;
+        }
         return friendService.friendSearchList(username);
     }
 
@@ -39,7 +63,7 @@ public class FriendController {
         HttpSession session = request.getSession();
         String name = (String) session.getAttribute("username");
         String role = (String) session.getAttribute("role");
-        if (StringUtils.isEmpty(name) || "1".equals(role)) {
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
             JSONObject jsonObject = new JSONObject();
             jsonObject.put("login", 0);
             return jsonObject;
@@ -51,38 +75,90 @@ public class FriendController {
 
     @RequestMapping(path = "/responseList")
     @ResponseBody
-    public JSONArray responseList(@RequestParam String username){
+    public JSONArray responseList(HttpServletRequest request, @RequestParam String username){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
+            JSONObject jsonObject = new JSONObject();
+            JSONArray jsonArray = new JSONArray();
+            jsonObject.put("login", 0);
+            jsonArray.add(jsonObject);
+            return jsonArray;
+        }
         return friendService.responseList(username);
     }
 
     @RequestMapping(path = "/requestList")
     @ResponseBody
-    public JSONArray requestList(@RequestParam String username){
+    public JSONArray requestList(HttpServletRequest request, @RequestParam String username){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
+            JSONObject jsonObject = new JSONObject();
+            JSONArray jsonArray = new JSONArray();
+            jsonObject.put("login", 0);
+            jsonArray.add(jsonObject);
+            return jsonArray;
+        }
         return friendService.requestList(username);
     }
 
     @RequestMapping(path = "/sendFriendRequest")
     @ResponseBody
-    public JSONObject sendFriendRequest(@RequestParam String username, @RequestParam String friend,
+    public JSONObject sendFriendRequest(HttpServletRequest request, @RequestParam String username, @RequestParam String friend,
                                         @RequestParam String content){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        }
         return friendService.sendFriendRequest(username, friend, content);
     }
 
     @RequestMapping(path = "/acceptRequest")
     @ResponseBody
-    public JSONObject acceptRequest(@RequestParam Long id){
+    public JSONObject acceptRequest(HttpServletRequest request, @RequestParam Long id){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role)) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        }
         return friendService.acceptRequest(id);
     }
 
     @RequestMapping(path = "/rejectRequest")
     @ResponseBody
-    public JSONObject rejectRequest(@RequestParam Long id){
+    public JSONObject rejectRequest(HttpServletRequest request, @RequestParam Long id){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role)) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        }
         return friendService.rejectRequest(id);
     }
 
     @RequestMapping(path = "/deleteFriend")
     @ResponseBody
-    public JSONObject deleteFriend(@RequestParam String username, @RequestParam String friend){
+    public JSONObject deleteFriend(HttpServletRequest request, @RequestParam String username, @RequestParam String friend){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || "1".equals(role) || !(name.equals(username))) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        }
         return friendService.deleteFriend(username, friend);
     }
